feat: expand CLI tooling, harden core subsystems, redesign TUI screens, add thread views
Broad enhancement across the MCP Agent Mail Rust stack (49 files,
+1496/-432 lines) covering CLI infrastructure, core reliability,
database layer, TUI interface, and end-to-end testing:

CLI improvements (mcp-agent-mail-cli):
- context.rs: refine execution context management for e2e test isolation
- e2e_artifacts.rs: improve test artifact collection and reporting
- e2e_runner.rs: enhance test runner with better failure diagnostics
- golden.rs: update golden test infrastructure for format evolution
- lib.rs: expand CLI entry points and command routing

Core hardening (mcp-agent-mail-core):
- bocpd.rs: refine Bayesian online changepoint detection for anomaly
  alerting on message flow patterns
- conformal.rs: improve conformal prediction intervals for latency
  estimation and SLA monitoring
- disk.rs: enhance on-disk persistence layer reliability
- evidence_ledger.rs: strengthen append-only evidence logging
- flake_triage.rs: improve flaky test detection and quarantine logic
- intern.rs: optimize string interning for reduced memory pressure
- lock_order.rs: refine deadlock-free lock ordering protocol
- proptest_generators.rs: expand property-based test generators
- setup.rs: streamline initialization and configuration

Database layer (mcp-agent-mail-db):
- integrity.rs: add consistency verification checks
- pool.rs: +93 lines expanding connection pool with health monitoring
- s3fifo.rs: tune S3-FIFO eviction cache for agent mail workloads
- search_service.rs: improve full-text search query execution
- tracking.rs: enhance message delivery tracking

TUI redesign (mcp-agent-mail-server):
- New thread_detail.rs: dedicated thread detail view with message
  threading visualization and inline reply
- New thread_list.rs: thread list view with sorting, filtering, and
  keyboard navigation
- tui_app.rs, tui_events.rs, tui_focus.rs: refactor application shell
  for new screen navigation model
- tui_macro.rs: +61 lines reworking the macro execution TUI
- tui_poller.rs: optimize background data refresh intervals
- Analytics, dashboard, explorer, messages, reservations, search,
  threads, timeline screens: incremental UX improvements
- tool_metrics.rs: refine tool-level performance instrumentation
- frame_bench.rs: update frame rendering benchmarks

Scripts and testing:
- New scripts/preflight.sh: pre-flight validation for deployment
- Update e2e test scripts: console, HTTP, MCP API parity, serve,
  TUI startup, TUI WASM
- Update TUI v3 interaction e2e test
- Refresh Cargo.lock for dependency changes
- Update beads issue tracker

+++ crates/mcp-agent-mail-core/src/conformal.rs
@@ -525,19 +525,20 @@ mod tests {
         );
     }
 
-    /// Exactly MIN_CALIBRATION observations: interval is available.
+    /// Exactly `MIN_CALIBRATION` observations: interval is available.
     #[test]
     fn conformal_exactly_min_calibration() {
         let mut predictor = ConformalPredictor::new(100, 0.90);
         for i in 0..MIN_CALIBRATION {
-            predictor.observe(f64::from(i as i32));
+            #[allow(clippy::cast_precision_loss)] // MIN_CALIBRATION is 30, well within f64 mantissa
+            predictor.observe(i as f64);
         }
         assert_eq!(predictor.calibration_size(), MIN_CALIBRATION);
         let interval = predictor.predict().unwrap();
         assert_eq!(interval.calibration_size, MIN_CALIBRATION);
     }
 
-    /// Window size of exactly 1 never produces a prediction (< MIN_CALIBRATION).
+    /// Window size of exactly 1 never produces a prediction (< `MIN_CALIBRATION`).
     #[test]
     fn conformal_tiny_window() {
         let mut predictor = ConformalPredictor::new(1, 0.90);
